requirements.txt
parsel==1.6.0
PyMySQL==1.0.2
Flask==1.1.2
Flask-Cors==3.0.10
PyYAML==5.3.1